Allow identifying DFU devices by serial number; refactoring
